+++ src/main/java/baseball/constant/ConstValue.java
@@ -1,10 +1,12 @@
-package baseball;
+package baseball.constant;
 
 public class ConstValue {
-	public static final int INITIAL_VALUE = 0;
 	public static final int MIN_NUMBER = 1;
 	public static final int MAX_NUMBER = 9;
 	public static final int NUMBER_LENGTH = 3;
+
+	public static final int INITIAL_VALUE = 0;
+	public static final int MAX_COUNT = 3;
 
 	public static final String RESTART = "1";
 	public static final String END_GAME = "2";
